// packages/api-client/src/generated/models/problemDetails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * GigPlanner.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ProblemDetails {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  title?: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  status?: number | string | null;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  instance?: string | null;
}